Convert Venmo tests to unit tests
