+++ Braintree/src/test/java/com/braintreepayments/api/internal/SignatureVerificationUnitTestUtils.java
@@ -1,6 +1,6 @@
 package com.braintreepayments.api.internal;
 
-public class SignatureVerificationTestUtils {
+public class SignatureVerificationUnitTestUtils {
 
     public static void disableSignatureVerification() {
         SignatureVerification.sEnableSignatureVerification = false;
